Added PR review changes , test coverage and new unit tests
