+++ README.md
@@ -1,8 +1,3 @@
-# Serice to expose the Mojaloop settlements API 
+# central-settlement
 
-The service is created using the hapi-openapi plugin.
-The current Swagger API documentation can be found [here](./config/swagger.json)
-The API is still only a design and the current code is the boilerplate for further development.
-
-This is part of the Mojaloop project.
-
+Swagger api [location](./config/swagger.json)
