+++ data/settlements/{Id}.js
@@ -1,48 +1,16 @@
-/*****
- License
- --------------
- Copyright © 2017 Bill & Melinda Gates Foundation
- The Mojaloop files are made available by the Bill & Melinda Gates Foundation under the Apache License, Version 2.0 (the "License") and you may not use these files except in compliance with the License. You may obtain a copy of the License at
-
- http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, the Mojaloop files are distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
- Contributors
- --------------
- This is the official list of the Mojaloop project contributors for this file.
- Names of the original copyright holders (individuals or organizations)
- should be listed with a '*' in the first column. People who have
- contributed from an organization can be listed under the organization
- that actually holds the copyright for their contributions (see the
- Gates Foundation organization for an example). Those individuals should have
- their names indented and be marked with a '-'. Email address can be added
- optionally within square brackets <email>.
-
- * Gates Foundation
- - Name Surname <name.surname@gatesfoundation.com>
-
- * Valentin Genev <valentin.genev@modusbox.com>
- * Deon Botha <deon.botha@modusbox.com>
- * Rajiv Mothilal <rajiv.mothilal@modusbox.com>
- * Miguel de Barros <miguel.debarros@modusbox.com>
-
- --------------
- ******/
 'use strict';
-
 var Mockgen = require('../mockgen.js');
 /**
- * Operations on /settlements/{Id}
+ * Operations on /settlements/{id}
  */
 module.exports = {
     /**
      * summary: Returns Settlement(s) as per parameters/filter criteria.
      * description: 
-     * parameters: Id, currency, Id
+     * parameters: id
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
-     * operationId: getSettlementsBySettlementParticipantCurrency
+     * operationId: getSettsettlementsById
      */
     get: {
         200: function (req, res, callback) {
@@ -51,7 +19,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'get',
                 response: '200'
             }, callback);
@@ -62,7 +30,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'get',
                 response: '400'
             }, callback);
@@ -73,7 +41,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'get',
                 response: '401'
             }, callback);
@@ -84,7 +52,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'get',
                 response: '404'
             }, callback);
@@ -95,7 +63,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'get',
                 response: '415'
             }, callback);
@@ -106,7 +74,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'get',
                 response: 'default'
             }, callback);
@@ -115,10 +83,10 @@ module.exports = {
     /**
      * summary: Acknowledegement of settlement by updating with Settlements Id.
      * description: 
-     * parameters: Id, settlementUpdatePayload
+     * parameters: id, settlementUpdatePayload
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
-     * operationId: updateSettlementBySettlementId
+     * operationId: updateSettlementById
      */
     put: {
         200: function (req, res, callback) {
@@ -127,7 +95,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'put',
                 response: '200'
             }, callback);
@@ -138,7 +106,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'put',
                 response: '400'
             }, callback);
@@ -149,7 +117,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'put',
                 response: '401'
             }, callback);
@@ -160,7 +128,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'put',
                 response: '404'
             }, callback);
@@ -171,7 +139,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'put',
                 response: '415'
             }, callback);
@@ -182,7 +150,7 @@ module.exports = {
              * Replace this by actual data for the api.
              */
             Mockgen().responses({
-                path: '/settlements/{Id}',
+                path: '/settlements/{id}',
                 operation: 'put',
                 response: 'default'
             }, callback);
