+++ handlers/settlementWindows.js
@@ -32,9 +32,8 @@
 
 'use strict';
 
-const Boom = require('boom');
 const dataAccess = require('../data/settlementWindows');
-const Logger = require('@mojaloop/central-services-shared').Logger
+const Boom = require('boom');
 
 /**
  * Operations on /settlementWindows
@@ -42,8 +41,8 @@ const Logger = require('@mojaloop/central-services-shared').Logger
 module.exports = {
     /**
      * summary: Returns a Settlement Window(s) as per parameter(s).
-     * description:
-     * parameters: id, state, fromDateTime, toDateTime
+     * description: 
+     * parameters: participantId, state, fromDateTime, toDateTime
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
      */
