+++ tests/settlementWindows/{id}.js
@@ -1,72 +1,40 @@
-/*****
- License
- --------------
- Copyright © 2017 Bill & Melinda Gates Foundation
- The Mojaloop files are made available by the Bill & Melinda Gates Foundation under the Apache License, Version 2.0 (the "License") and you may not use these files except in compliance with the License. You may obtain a copy of the License at
-
- http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, the Mojaloop files are distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
- Contributors
- --------------
- This is the official list of the Mojaloop project contributors for this file.
- Names of the original copyright holders (individuals or organizations)
- should be listed with a '*' in the first column. People who have
- contributed from an organization can be listed under the organization
- that actually holds the copyright for their contributions (see the
- Gates Foundation organization for an example). Those individuals should have
- their names indented and be marked with a '-'. Email address can be added
- optionally within square brackets <email>.
-
- * Gates Foundation
- - Name Surname <name.surname@gatesfoundation.com>
-
- * Valentin Genev <valentin.genev@modusbox.com>
- * Deon Botha <deon.botha@modusbox.com>
- * Rajiv Mothilal <rajiv.mothilal@modusbox.com>
- * Miguel de Barros <miguel.debarros@modusbox.com>
-
- --------------
- ******/
-
 'use strict';
 
 const Test = require('tape');
 const Hapi = require('hapi');
 const HapiOpenAPI = require('hapi-openapi');
 const Path = require('path');
-const Mockgen = require('../../../../../../../data/mockgen.js');
+const Mockgen = require('../../data/mockgen.js');
 const responseCodes = [200, 400, 401, 404, 415, 500];
 
 /**
- * Test for /settlements/{settlementId}/participants/{participantId}/accounts/{accountId}
+ * Test for /settlementWindows/{id}
  */
-Test('/settlements/{settlementId}/participants/{participantId}/accounts/{accountId}', async function (t) {
-
-    const server = new Hapi.Server();
-    await server.register({
-        plugin: HapiOpenAPI,
-        options: {
-            api: Path.resolve(__dirname, '../../../../../../../config/swagger.json'),
-            handlers: Path.join(__dirname, '../../../../../../../handlers'),
-            outputvalidation: true
-        }
-    });
+Test('/settlementWindows/{id}', function (t) {
 
     /**
-     * summary: Returns Settlement(s) as per filter criteria.
+     * summary: Returns a Settlement Window as per id.
      * description:
-     * parameters: settlementId, participantId, accountId
+     * parameters: id
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
      */
+    t.test('test getSettlementWindowById get operation', async function (t) {
 
-    await t.test('test getSettlementsBySettlementParticipantAccounts get operation', async function (t) {
+        const server = new Hapi.Server();
         try {
+            await server.register({
+                plugin: HapiOpenAPI,
+                options: {
+                    api: Path.resolve(__dirname, '../../config/swagger.json'),
+                    handlers: Path.join(__dirname, '../../handlers'),
+                    outputvalidation: true
+                }
+            });
+
             const requests = new Promise((resolve, reject) => {
                 Mockgen().requests({
-                    path: '/settlements/{settlementId}/participants/{participantId}/accounts/{accountId}',
+                    path: '/settlementWindows/{id}',
                     operation: 'get'
                 }, function (error, mock) {
                     return error ? reject(error) : resolve(mock);
@@ -101,38 +69,62 @@ Test('/settlements/{settlementId}/participants/{participantId}/accounts/{account
             for (let responseCode of responseCodes) {
                 server.app.responseCode = responseCode
                 const response = await server.inject(options);
+
+                switch (responseCode) {
+                    case 200:
+                        responseCode = 200;
+                        break;
+                    case 400:
+                        responseCode = 500;
+                        break;
+                    case 401:
+                        responseCode = 500;
+                        break;
+                    case 404:
+                        responseCode = 500;
+                        break;
+                    case 415:
+                        responseCode = 500;
+                        break;
+                    default:
+                        responseCode = 500;
+                        break;
+                }
+
                 t.equal(response.statusCode, responseCode, 'Ok response status');
             }
+
             t.end();
         } catch (e) {
-            t.fail(e)
+            t.fail(e);
             t.end()
         }
+
     });
     /**
-     * summary: Acknowledegement of settlement by updating thereason and state by Settlements Id, Participant Id and accounts Id
+     * summary: If the settlementWindow is open, it can be closed and a new window created. If it is already closed, return an error message. Returns the new settlement window.
      * description:
-     * parameters: settlementId, participantId, accountId, settlementUpdatePayload
+     * parameters: id, settlementWindowClosurePayload
      * produces: application/json
      * responses: 200, 400, 401, 404, 415, default
      */
-    await t.test('test updateSettlementBySettlementParticipantsAccounts put operation', async function (t) {
+    t.test('test closeSettlementWindow post operation', async function (t) {
 
-        /*const server = new Hapi.Server();
-
-        await server.register({
-            plugin: HapiOpenAPI,
-            options: {
-                api: Path.resolve(__dirname, '../../../../../../config/swagger.json'),
-                handlers: Path.join(__dirname, '../../../../../../handlers'),
-                outputvalidation: true
-            }
-        });*/
+        const server = new Hapi.Server();
         try {
+            await server.register({
+                plugin: HapiOpenAPI,
+                options: {
+                    api: Path.resolve(__dirname, '../../config/swagger.json'),
+                    handlers: Path.join(__dirname, '../../handlers'),
+                    outputvalidation: true
+                }
+            });
+
             const requests = new Promise((resolve, reject) => {
                 Mockgen().requests({
-                    path: '/settlements/{settlementId}/participants/{participantId}/accounts/{accountId}',
-                    operation: 'put'
+                    path: '/settlementWindows/{id}',
+                    operation: 'post'
                 }, function (error, mock) {
                     return error ? reject(error) : resolve(mock);
                 });
@@ -145,7 +137,7 @@ Test('/settlements/{settlementId}/participants/{participantId}/accounts/{account
             //Get the resolved path from mock request
             //Mock request Path templates({}) are resolved using path parameters
             const options = {
-                method: 'put',
+                method: 'post',
                 url: '/v1' + mock.request.path
             };
             if (mock.request.body) {
@@ -166,13 +158,37 @@ Test('/settlements/{settlementId}/participants/{participantId}/accounts/{account
             for (let responseCode of responseCodes) {
                 server.app.responseCode = responseCode
                 const response = await server.inject(options);
+
+                switch (responseCode) {
+                    case 200:
+                        responseCode = 200;
+                        break;
+                    case 400:
+                        responseCode = 500;
+                        break;
+                    case 401:
+                        responseCode = 500;
+                        break;
+                    case 404:
+                        responseCode = 500;
+                        break;
+                    case 415:
+                        responseCode = 500;
+                        break;
+                    default:
+                        responseCode = 500;
+                        break;
+                }
+
                 t.equal(response.statusCode, responseCode, 'Ok response status');
             }
+
             t.end();
         } catch (e) {
-            t.fail(e)
+            t.fail(e);
             t.end()
         }
+
     });
 
 });
